move the MultiDrawable class to the new map package in the ui package
